requirements.txt
openenv-core>=0.2.3
pydantic>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
openai>=1.30.0
datasets>=2.19.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
groq>=0.4.0